Release 1.0.1 - Fix README.md. Move to project specific package. Use real Discord ID for author. Fix manifest. Remove kotlin files.
