+++ FxTwitterAlways/build.gradle
@@ -1,9 +1,10 @@
-version = "1.0.0" // Plugin version. Increment this to trigger the updater
+version = "1.0.1" // Plugin version. Increment this to trigger the updater
 description = "Always send twitter links with your chosen twitter fix domain. Bonus: lose the query params too" // Plugin description that will be shown to user
 aliucord {
     // Changelog of your plugin
     changelog = """
-        Initial version.
+        1.0.0 Initial version.
+        1.0.1 Fix manifest etc. to hopefully get the PluginDownloader working.
     """
     // Image or Gif that will be shown at the top of your changelog page
     // changelogMedia.set("https://cool.png")
