+++ FxTwitterAlways/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="youImportedTheWrongR" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.github.tonka92.fxtwitteralways" />
