--- build.gradle2.kts
@@ -1,75 +1,0 @@
-import com.aliucord.gradle.AliucordExtension
-import com.android.build.gradle.BaseExtension
-
-buildscript {
-    repositories {
-        google()
-        mavenCentral()
-        // Aliucords Maven repo which contains our tools and dependencies
-        maven("https://maven.aliucord.com/snapshots")
-        // Shitpack which still contains some Aliucord dependencies for now. TODO: Remove
-        maven("https://jitpack.io")
-    }
-
-    dependencies {
-        classpath("com.android.tools.build:gradle:7.0.4")
-        // Aliucord gradle plugin which makes everything work and builds plugins
-        classpath("com.aliucord:gradle:main-SNAPSHOT")
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-        maven("https://maven.aliucord.com/snapshots")
-    }
-}
-
-fun Project.aliucord(configuration: AliucordExtension.() -> Unit) = extensions.getByName<AliucordExtension>("aliucord").configuration()
-
-fun Project.android(configuration: BaseExtension.() -> Unit) = extensions.getByName<BaseExtension>("android").configuration()
-
-subprojects {
-    apply(plugin = "com.android.library")
-    apply(plugin = "com.aliucord.gradle")
-
-    // Fill out with your info
-    aliucord {
-        author("DISCORD USERNAME", 123456789L)
-        updateUrl.set("https://raw.githubusercontent.com/USERNAME/REPONAME/builds/updater.json")
-        buildUrl.set("https://raw.githubusercontent.com/USERNAME/REPONAME/builds/%s.zip")
-    }
-
-    android {
-        compileSdkVersion(31)
-
-        defaultConfig {
-            minSdk = 24
-            targetSdk = 31
-        }
-
-        compileOptions {
-            sourceCompatibility = JavaVersion.VERSION_11
-            targetCompatibility = JavaVersion.VERSION_11
-        }
-
-    }
-
-    dependencies {
-        val discord by configurations
-        val implementation by configurations
-
-        // Stubs for all Discord classes
-        discord("com.discord:discord:aliucord-SNAPSHOT")
-        implementation("com.aliucord:Aliucord:main-SNAPSHOT")
-
-        implementation("androidx.appcompat:appcompat:1.4.0")
-        implementation("com.google.android.material:material:1.4.0")
-        implementation("androidx.constraintlayout:constraintlayout:2.1.2")
-    }
-}
-
-task<Delete>("clean") {
-    delete(rootProject.buildDir)
-}
